VisionLink-main/requirements.txt
depthai==2.21.2.0
keras==2.12.0
numpy==1.23.5
opencv_python==4.7.0.72
# paddle==1.0.2
Pillow==9.5.0
pyclipper==1.3.0.post4
pydantic==1.10.7
PyYAML==6.0
PyYAML==6.0
Shapely==2.0.1
#tensorflow==2.12.0
